### 2025-02-20 #7829 @editorialbot
[PRE REVIEW]: The GeoRegions Ecosystem: A Simple Julia Ecosystem for Gridded and other Spatial Data in Earth Observation
**Submitting author:** @natgeo-wong (<a href="http://orcid.org/0000-0003-4744-3661">Nathanael Wong</a>)
**Repository:** https://github.com/GeoRegionsEcosystem/GeoRegions.jl
**Branch with paper.md** (empty if default branch):
**Version:** v8.0.1
