### 2025-02-20 #7829 @natgeo-wong
Hi @kthyng thanks so much! The repositories are:

https://github.com/GeoRegionsEcosystem/GeoRegions.jl
https://github.com/GeoRegionsEcosystem/RegionGrids.jl
https://github.com/GeoRegionsEcosystem/LandSea.jl
