### 2025-02-26 5f8ebad
Modify bounds checking a bit, bump to v8.0.3
