+++ dev/assets/tutorials_using_isin.md.BJgSoFUr.js
@@ -1,4 +1,4 @@
-import{_ as i,c as a,o as h,aj as n}from"./chunks/framework.97ulaogc.js";const k="/GeoRegions.jl/dev/assets/gfvyabi.laFYj86U.png",t="/GeoRegions.jl/dev/assets/kxhfkhc.aiS5MYw1.png",o=JSON.parse('{"title":"Is it in a GeoRegion?","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/isin.md","filePath":"tutorials/using/isin.md","lastUpdated":null}'),l={name:"tutorials/using/isin.md"};function p(e,s,E,d,g,r){return h(),a("div",null,s[0]||(s[0]=[n(`<h1 id="Is-it-in-a-GeoRegion?" tabindex="-1">Is it in a GeoRegion? <a class="header-anchor" href="#Is-it-in-a-GeoRegion?" aria-label="Permalink to &quot;Is it in a GeoRegion? {#Is-it-in-a-GeoRegion?}&quot;">​</a></h1><p>When dealing with geographic data, we often wish to check if a point or region is inside another region. GeoRegions.jl allows you to perform this check easily with the function <code>isinGeoRegion</code>.</p><div class="tip custom-block"><p class="custom-block-title">Point2 Type</p><p>We use the <code>Point2</code> Type from the package <a href="https://github.com/JuliaGeometry/GeometryBasics.jl" target="_blank" rel="noreferrer">GeometryBasics.jl</a>, which is reexported by GeoRegions.jl, as an easy way to denote points. This also allows us to use the package <a href="https://github.com/JuliaGeo/GeometryOps.jl" target="_blank" rel="noreferrer">GeometryOps.jl</a> to determine if a point is within a GeoRegion.</p></div><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
[…]
