+++ dev/assets/api_createaddreadrm.md.E4LKeZWI.lean.js
@@ -1,1 +1,0 @@
-import{_ as o,C as l,c as h,o as p,j as i,a as e,aj as n,G as t}from"./chunks/framework.97ulaogc.js";const A=JSON.parse('{"title":"API for Creating, Reading, Adding and Removing GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"api/createaddreadrm.md","filePath":"api/createaddreadrm.md","lastUpdated":null}'),r={name:"api/createaddreadrm.md"},d={class:"jldocstring custom-block",open:""},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""};function u(m,s,F,C,R,f){const a=l("Badge");return p(),h("div",null,[s[18]||(s[18]=i("h1",{id:"API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions",tabindex:"-1"},[e("API for Creating, Reading, Adding and Removing GeoRegions "),i("a",{class:"header-anchor",href:"#API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions","aria-label":'Permalink to "API for Creating, Reading, Adding and Removing GeoRegions {#API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions}"'},"​")],-1)),s[19]||(s[19]=i("h2",{id:"Creating-GeoRegions",tabindex:"-1"},[e("Creating GeoRegions "),i("a",{class:"header-anchor",href:"#Creating-GeoRegions","aria-label":'Permalink to "Creating GeoRegions {#Creating-GeoRegions}"'},"​")],-1)),i("details",d,[i("summary",null,[s[0]||(s[0]=i("a",{id:"GeoRegions.GeoRegion-Tuple{Vector{<:Real}, Vector{<:Real}}",href:"#GeoRegions.GeoRegion-Tuple{Vector{<:Real}, Vector{<:Real}}"},[i("span",{class:"jlbinding"},"GeoRegions.GeoRegion")],-1)),s[1]||(s[1]=e()),t(a,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[2]||(s[2]=n("",9))]),s[20]||(s[20]=i("h2",{id:"Retrieving-GeoRegions",tabindex:"-1"},[e("Retrieving GeoRegions "),i("a",{class:"header-anchor",href:"#Retrieving-GeoRegions","aria-label":'Permalink to "Retrieving GeoRegions {#Retrieving-GeoRegions}"'},"​")],-1)),i("details",k,[i("summary",null,[s[3]||(s[3]=i("a",{id:"GeoRegions.GeoRegion-Tuple{AbstractString}",href:"#GeoRegions.GeoRegion-Tuple{AbstractString}"},[i("span",{class:"jlbinding"},"GeoRegions.GeoRegion")],-1)),s[4]||(s[4]=e()),t(a,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[5]||(s[5]=n("",9))]),s[21]||(s[21]=i("h2",{id:"Adding-Custom-GeoRegions",tabindex:"-1"},[e("Adding Custom GeoRegions "),i("a",{class:"header-anchor",href:"#Adding-Custom-GeoRegions","aria-label":'Permalink to "Adding Custom GeoRegions {#Adding-Custom-GeoRegions}"'},"​")],-1)),i("details",g,[i("summary",null,[s[6]||(s[6]=i("a",{id:"GeoRegions.add",
[…]
