+++ dev/assets/tutorials_using_derotate.md.Cblkhddy.js
@@ -1,4 +1,4 @@
-import{_ as i,c as a,o as n,aj as h}from"./chunks/framework.97ulaogc.js";const t="/GeoRegions.jl/dev/assets/ibcvupu.CRTH70aj.png",k="/GeoRegions.jl/dev/assets/fapdomg.BaDECfsY.png",y=JSON.parse('{"title":"Derotating Coordinates around a given GeoRegion","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/derotate.md","filePath":"tutorials/using/derotate.md","lastUpdated":null}'),l={name:"tutorials/using/derotate.md"};function p(e,s,E,r,d,g){return n(),a("div",null,s[0]||(s[0]=[h(`<h1 id="Derotating-Coordinates-around-a-given-GeoRegion" tabindex="-1">Derotating Coordinates around a given GeoRegion <a class="header-anchor" href="#Derotating-Coordinates-around-a-given-GeoRegion" aria-label="Permalink to &quot;Derotating Coordinates around a given GeoRegion {#Derotating-Coordinates-around-a-given-GeoRegion}&quot;">​</a></h1><p>Using the rotation and centroid properties of a GeoRegion, we are able to perform derotation of coordinates about the centroid of a given GeoRegion, allowing us to project longitude/latitude coordinates and even data into a x/y-cartesian coordinate system.</p><p>Let us now setup the example</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
+import{_ as i,c as a,o as n,aj as h}from"./chunks/framework.97ulaogc.js";const t="/GeoRegions.jl/dev/assets/riectjj.CRTH70aj.png",k="/GeoRegions.jl/dev/assets/mrlxcop.BaDECfsY.png",y=JSON.parse('{"title":"Derotating Coordinates around a given GeoRegion","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/derotate.md","filePath":"tutorials/using/derotate.md","lastUpdated":null}'),l={name:"tutorials/using/derotate.md"};function p(e,s,E,r,d,g){return n(),a("div",null,s[0]||(s[0]=[h(`<h1 id="Derotating-Coordinates-around-a-given-GeoRegion" tabindex="-1">Derotating Coordinates around a given GeoRegion <a class="header-anchor" href="#Derotating-Coordinates-around-a-given-GeoRegion" aria-label="Permalink to &quot;Derotating Coordinates around a given GeoRegion {#Derotating-Coordinates-around-a-given-GeoRegion}&quot;">​</a></h1><p>Using the rotation and centroid properties of a GeoRegion, we are able to perform derotation of coordinates about the centroid of a given GeoRegion, allowing us to project longitude/latitude coordinates and even data into a x/y-cartesian coordinate system.</p><p>Let us now setup the example</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
[…]

[…]
