+++ v8.0.3/assets/api_isinonequal.md.Ch3Cdyxn.lean.js
@@ -1,0 +1,1 @@
+import{_ as o,C as l,c as h,o as p,aj as a,j as i,a as e,G as t}from"./chunks/framework.DbkP9AJj.js";const T=JSON.parse('{"title":"API for Is it in/on/equal to a GeoRegion?","description":"","frontmatter":{},"headers":[],"relativePath":"api/isinonequal.md","filePath":"api/isinonequal.md","lastUpdated":null}'),d={name:"api/isinonequal.md"},k={class:"jldocstring custom-block",open:""},r={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},f={class:"jldocstring custom-block",open:""},b={class:"jldocstring custom-block",open:""},F={class:"jldocstring custom-block",open:""};function C(R,s,G,A,m,D){const n=l("Badge");return p(),h("div",null,[s[30]||(s[30]=a("",3)),i("details",k,[i("summary",null,[s[0]||(s[0]=i("a",{id:"Base.in-Tuple{Point2{<:Real}, GeoRegion}",href:"#Base.in-Tuple{Point2{<:Real}, GeoRegion}"},[i("span",{class:"jlbinding"},"Base.in")],-1)),s[1]||(s[1]=e()),t(n,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[2]||(s[2]=a("",9))]),s[31]||(s[31]=i("h3",{id:"Is-GeoRegion-In-a-GeoRegion",tabindex:"-1"},[e("Is GeoRegion In a GeoRegion "),i("a",{class:"header-anchor",href:"#Is-GeoRegion-In-a-GeoRegion","aria-label":'Permalink to "Is GeoRegion In a GeoRegion {#Is-GeoRegion-In-a-GeoRegion}"'},"​")],-1)),i("details",r,[i("summary",null,[s[3]||(s[3]=i("a",{id:"Base.in-Tuple{GeoRegion, GeoRegion}",href:"#Base.in-Tuple{GeoRegion, GeoRegion}"},[i("span",{class:"jlbinding"},"Base.in")],-1)),s[4]||(s[4]=e()),t(n,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[5]||(s[5]=a("",9))]),s[32]||(s[32]=i("h2",{id:"on",tabindex:"-1"},[e("On "),i("a",{class:"header-anchor",href:"#on","aria-label":'Permalink to "On"'},"​")],-1)),s[33]||(s[33]=i("h3",{id:"Is-Point-On-a-GeoRegion",tabindex:"-1"},[e("Is Point On a GeoRegion "),i("a",{class:"header-anchor",href:"#Is-Point-On-a-GeoRegion","aria-label":'Permalink to "Is Point On a GeoRegion {#Is-Point-On-a-GeoRegion}"'},"​")],-1)),i("details",g,[i("summary",null,[s[6]||(s[6]=i("a",{id:"GeoRegions.on-Tuple{Point2{<:Real}, GeoRegion}",href:"#GeoRegions.on-Tuple{Point2{<:Real}, GeoRegion}"},[i("span",{class:"jlbinding"},"GeoRegions.on")],-1)),s[7]||(s[7]=e()),t(n,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[8]||(s[8]=a("",9))]),s[34]||(s[
[…]
