+++ v8.0.3/assets/basics_tables.md.BXuwlKO_.lean.js
@@ -1,0 +1,1 @@
+import{_ as a,c as e,o as i,aj as n}from"./chunks/framework.DbkP9AJj.js";const g=JSON.parse('{"title":"Tabular Listings of GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/tables.md","filePath":"basics/tables.md","lastUpdated":null}'),t={name:"basics/tables.md"};function o(l,s,p,r,h,d){return i(),e("div",null,s[0]||(s[0]=[n("",10)]))}const k=a(t,[["render",o]]);export{g as __pageData,k as default};
