### 2025-03-06 a784bc5
Flush logging when needed
