+++ dev/assets/api_project.md.C6R-T3ac.lean.js
@@ -1,1 +1,0 @@
-import{_ as l,C as o,c as p,o as h,j as i,a as e,aj as t,G as n}from"./chunks/framework.97ulaogc.js";const j=JSON.parse('{"title":"API for Project-Handling in GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"api/project.md","filePath":"api/project.md","lastUpdated":null}'),d={name:"api/project.md"},r={class:"jldocstring custom-block",open:""},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""};function y(E,s,u,f,F,b){const a=o("Badge");return h(),p("div",null,[s[12]||(s[12]=i("h1",{id:"API-for-Project-Handling-in-GeoRegions.jl",tabindex:"-1"},[e("API for Project-Handling in GeoRegions.jl "),i("a",{class:"header-anchor",href:"#API-for-Project-Handling-in-GeoRegions.jl","aria-label":'Permalink to "API for Project-Handling in GeoRegions.jl {#API-for-Project-Handling-in-GeoRegions.jl}"'},"​")],-1)),s[13]||(s[13]=i("h2",{id:"Setting-up-GeoRegions.jl-in-a-Project",tabindex:"-1"},[e("Setting up GeoRegions.jl in a Project "),i("a",{class:"header-anchor",href:"#Setting-up-GeoRegions.jl-in-a-Project","aria-label":'Permalink to "Setting up GeoRegions.jl in a Project {#Setting-up-GeoRegions.jl-in-a-Project}"'},"​")],-1)),i("details",r,[i("summary",null,[s[0]||(s[0]=i("a",{id:"GeoRegions.setupGeoRegions",href:"#GeoRegions.setupGeoRegions"},[i("span",{class:"jlbinding"},"GeoRegions.setupGeoRegions")],-1)),s[1]||(s[1]=e()),n(a,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),s[2]||(s[2]=t("",5))]),i("details",k,[i("summary",null,[s[3]||(s[3]=i("a",{id:"GeoRegions.deleteGeoRegions",href:"#GeoRegions.deleteGeoRegions"},[i("span",{class:"jlbinding"},"GeoRegions.deleteGeoRegions")],-1)),s[4]||(s[4]=e()),n(a,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),s[5]||(s[5]=t("",5))]),s[14]||(s[14]=i("h2",{id:"Handling-custom-lists-of-GeoRegions.jl",tabindex:"-1"},[e("Handling custom lists of GeoRegions.jl "),i("a",{class:"header-anchor",href:"#Handling-custom-lists-of-GeoRegions.jl","aria-label":'Permalink to "Handling custom lists of GeoRegions.jl {#Handling-custom-lists-of-GeoRegions.jl}"'},"​")],-1)),i("details",g,[i("summary",null,[s[6]||(s[6]=i("a",{id:"GeoRegions.loadGeoRegions",href:"#GeoRegions.loadGeoRegions"},[i("span",{class:"jlbinding"},"GeoRegions.loadGeoRegions")],-1)),s[7]||(s[7]=e()),n(a,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),s[8]||(s[8]=t("",7))]),i("details",c,[i("summary",null,[s[9]||(s[9]=i("a",{id:"GeoRegions.addGeoRegions",href:"#GeoRegions.addGeoRegions"},[i("span",{class:"jlbinding"},"GeoRegions.addGeoRegions")],-1)),
[…]
