+++ dev/assets/basics_predefined_datasets.md.4HkkerOP.lean.js
@@ -1,1 +1,0 @@
-import{_ as s,c as a,o as n,aj as i}from"./chunks/framework.97ulaogc.js";const g=JSON.parse('{"title":"Predefined GeoRegions in GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"basics/predefined/datasets.md","filePath":"basics/predefined/datasets.md","lastUpdated":null}'),o={name:"basics/predefined/datasets.md"};function t(p,e,l,d,r,c){return n(),a("div",null,e[0]||(e[0]=[i("",21)]))}const R=s(o,[["render",t]]);export{g as __pageData,R as default};
