+++ dev/assets/api_project.md.F3ZyzT4S.js
@@ -1,0 +1,19 @@
+import{_ as o,C as p,c as h,o as d,j as i,a as t,aj as n,G as a,w as l}from"./chunks/framework.OnoNBi2O.js";const R=JSON.parse('{"title":"API for Project-Handling in GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"api/project.md","filePath":"api/project.md","lastUpdated":null}'),r={name:"api/project.md"},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""};function E(u,s,f,F,C,b){const e=p("Badge");return d(),h("div",null,[s[16]||(s[16]=i("h1",{id:"API-for-Project-Handling-in-GeoRegions.jl",tabindex:"-1"},[t("API for Project-Handling in GeoRegions.jl "),i("a",{class:"header-anchor",href:"#API-for-Project-Handling-in-GeoRegions.jl","aria-label":'Permalink to "API for Project-Handling in GeoRegions.jl {#API-for-Project-Handling-in-GeoRegions.jl}"'},"​")],-1)),s[17]||(s[17]=i("h2",{id:"Setting-up-GeoRegions.jl-in-a-Project",tabindex:"-1"},[t("Setting up GeoRegions.jl in a Project "),i("a",{class:"header-anchor",href:"#Setting-up-GeoRegions.jl-in-a-Project","aria-label":'Permalink to "Setting up GeoRegions.jl in a Project {#Setting-up-GeoRegions.jl-in-a-Project}"'},"​")],-1)),i("details",k,[i("summary",null,[s[0]||(s[0]=i("a",{id:"GeoRegions.setupGeoRegions",href:"#GeoRegions.setupGeoRegions"},[i("span",{class:"jlbinding"},"GeoRegions.setupGeoRegions")],-1)),s[1]||(s[1]=t()),a(e,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),s[3]||(s[3]=n(`<div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">setupGeoRegions</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(;</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    path </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> AbstractString</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;"> =</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> pwd</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(),</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    overwrite </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Bool</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;"> =</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> false</span></span>
[…]
