+++ v8.0.4/assets/basics_read.md.Be-efMnm.js
@@ -1,0 +1,10 @@
+import{_ as s,c as a,o as i,aj as n}from"./chunks/framework.vQw0R27l.js";const g=JSON.parse('{"title":"Reading Predefined GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/read.md","filePath":"basics/read.md","lastUpdated":null}'),o={name:"basics/read.md"};function t(p,e,l,d,r,h){return i(),a("div",null,e[0]||(e[0]=[n(`<h1 id="Reading-Predefined-GeoRegions" tabindex="-1">Reading Predefined <code>GeoRegion</code>s <a class="header-anchor" href="#Reading-Predefined-GeoRegions" aria-label="Permalink to &quot;Reading Predefined \`GeoRegion\`s {#Reading-Predefined-GeoRegions}&quot;">​</a></h1><p>The most basic functionality is to retrieve the information of a <code>GeoRegion</code>. There are several sets of predefined GeoRegions in GeoRegions.jl, listed <a href="/GeoRegions.jl/v8.0.4/basics/predefined/listall">here</a>.</p><p>You can read and retrieve the properties of a <code>GeoRegion</code> using the function <code>GeoRegion()</code> for a given <code>ID</code>, which is a <code>String</code> type.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">geo </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> GeoRegion</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(ID)</span></span></code></pre></div><p>For example, the default <code>GeoRegion</code> spans the whole globe. Its ID is <code>ID = &quot;GLB&quot;</code>.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">geo </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> GeoRegion</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;">&quot;GLB&quot;</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span></code></pre></div><div class="language- vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang"></span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span>The GeoRegion GLB has the following properties:</span></span>
+<span class="line"><span>    Region ID     (ID) : GLB</span></span>
+<span class="line"><span>    Parent ID    (pID) : GLB</span></span>
+<span class="line"><span>    Name        (name) : Globe</span></span>
[…]
