+++ v8.0.4/assets/tutorials_projects_addreadrm.md.CjO1ffQ2.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as n,aj as e}from"./chunks/framework.vQw0R27l.js";const t="/GeoRegions.jl/v8.0.4/assets/bzevqho.B0LZKJCn.png",E=JSON.parse('{"title":"Manipulation of User-Defined GeoRegions.jl for your Project","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/projects/addreadrm.md","filePath":"tutorials/projects/addreadrm.md","lastUpdated":null}'),h={name:"tutorials/projects/addreadrm.md"};function p(l,s,k,o,d,r){return n(),a("div",null,s[0]||(s[0]=[e("",78)]))}const c=i(h,[["render",p]]);export{E as __pageData,c as default};
