+++ v8.0.4/assets/basics_predefined_listall.md.JbVw-qUH.lean.js
@@ -1,0 +1,1 @@
+import{_ as a,c as n,o as i,aj as e}from"./chunks/framework.vQw0R27l.js";const d=JSON.parse('{"title":"Default List of Predefined GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/predefined/listall.md","filePath":"basics/predefined/listall.md","lastUpdated":null}'),p={name:"basics/predefined/listall.md"};function l(t,s,o,r,c,h){return i(),n("div",null,s[0]||(s[0]=[e("",22)]))}const R=a(p,[["render",l]]);export{d as __pageData,R as default};
