+++ v8.0.4/assets/basics_create.md.CpxrXu4z.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as n,aj as e}from"./chunks/framework.vQw0R27l.js";const g=JSON.parse('{"title":"Creating Custom GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/create.md","filePath":"basics/create.md","lastUpdated":null}'),t={name:"basics/create.md"};function p(l,s,h,k,o,r){return n(),a("div",null,s[0]||(s[0]=[e("",5)]))}const c=i(t,[["render",p]]);export{g as __pageData,c as default};
