+++ v8.0.4/assets/basics_read.md.Be-efMnm.lean.js
@@ -1,0 +1,1 @@
+import{_ as s,c as a,o as i,aj as n}from"./chunks/framework.vQw0R27l.js";const g=JSON.parse('{"title":"Reading Predefined GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/read.md","filePath":"basics/read.md","lastUpdated":null}'),o={name:"basics/read.md"};function t(p,e,l,d,r,h){return i(),a("div",null,e[0]||(e[0]=[n("",9)]))}const k=s(o,[["render",t]]);export{g as __pageData,k as default};

[…]
