### 2025-03-06 #7829 @kthyng
@natgeo-wong No I haven't had the time to get to your submission yet, but fear not — it's not holding it back from review anyway since it is about 20th in line in the inbox and we don't have any editors available at the moment

If you're keen to get going you can address the items above like mentioned in my comment — the missing DOIs, suggesting reviewers, shortening your paper to max 1000 words.
