### 2025-03-08 7d379f3
Only the style.css shall remain
