+++ dev/assets/tutorials_projects_backends.md.CU-cTKOy.js
@@ -1,16 +1,0 @@
-import{_ as i,c as a,o as e,aj as n}from"./chunks/framework.OnoNBi2O.js";const g=JSON.parse('{"title":"Backends for GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/projects/backends.md","filePath":"tutorials/projects/backends.md","lastUpdated":null}'),t={name:"tutorials/projects/backends.md"};function p(l,s,h,o,k,r){return e(),a("div",null,s[0]||(s[0]=[n(`<h1 id="Backends-for-GeoRegions.jl" tabindex="-1">Backends for GeoRegions.jl <a class="header-anchor" href="#Backends-for-GeoRegions.jl" aria-label="Permalink to &quot;Backends for GeoRegions.jl {#Backends-for-GeoRegions.jl}&quot;">​</a></h1><p>This page gives a brief summary of the backend process that GeoRegions.j uses to save <strong>manually</strong> defined custom <code>GeoRegion</code>s for projects.</p><h2 id="JSON-files" tabindex="-1">JSON files <a class="header-anchor" href="#JSON-files" aria-label="Permalink to &quot;JSON files {#JSON-files}&quot;">​</a></h2><p>All relevant information for a GeoRegion can be divided into two components:</p><ul><li><p>String descriptors that identify a GeoRegion and make each GeoRegion unique</p></li><li><p>Geometry, or a vector of Geometries, that describe the geographic region itself</p></li></ul><p>Ever since ≥v8, I have decided that JSON is the appropriate format to store this information. Therefore, each <code>GeoRegion</code> will have its own unique <code>.json</code> file, stored in the location <code>$path/.georegions/$ID.json</code>, where <code>path</code> is the directory of your project.</p><p>We use <a href="./backends">JSON3.jl</a> to write and read these <code>.json</code> files, as JSON3.jl is able to take a <code>struct</code> and parse it into a readable JSON string.</p><p>Let&#39;s show an example</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">geo </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> GeoRegion</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span></span>
[…]
