+++ dev/assets/tutorials_using_ison.md.8mwoT2pa.lean.js
@@ -1,1 +1,0 @@
-import{_ as i,c as a,o as n,aj as e}from"./chunks/framework.OnoNBi2O.js";const t="/GeoRegions.jl/dev/assets/duwnuyi.DK3TS_Gw.png",E=JSON.parse('{"title":"Is it on a GeoRegion Boundary?","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/ison.md","filePath":"tutorials/using/ison.md","lastUpdated":null}'),h={name:"tutorials/using/ison.md"};function p(l,s,k,o,d,r){return n(),a("div",null,s[0]||(s[0]=[e("",48)]))}const c=i(h,[["render",p]]);export{E as __pageData,c as default};
