+++ dev/assets/basics_read.md._f2OJ-xG.lean.js
@@ -1,1 +1,0 @@
-import{_ as s,c as a,o as i,aj as n}from"./chunks/framework.OnoNBi2O.js";const g=JSON.parse('{"title":"Reading Predefined GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/read.md","filePath":"basics/read.md","lastUpdated":null}'),o={name:"basics/read.md"};function t(p,e,d,l,r,h){return i(),a("div",null,e[0]||(e[0]=[n("",9)]))}const k=s(o,[["render",t]]);export{g as __pageData,k as default};
