+++ dev/assets/tutorials_using_derotate.md.BSUhc6k7.lean.js
@@ -1,1 +1,0 @@
-import{_ as i,c as a,o as n,aA as h}from"./chunks/framework.D3LA_rsn.js";const t="/GeoRegions.jl/dev/assets/nmqfffy.DkCBYHpQ.png",k="/GeoRegions.jl/dev/assets/ndzsprv.BfsmKH3z.png",y=JSON.parse('{"title":"Derotating Coordinates around a given GeoRegion","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/derotate.md","filePath":"tutorials/using/derotate.md","lastUpdated":null}'),l={name:"tutorials/using/derotate.md"};function p(e,s,E,r,d,g){return n(),a("div",null,s[0]||(s[0]=[h("",21)]))}const F=i(l,[["render",p]]);export{y as __pageData,F as default};
