+++ v8.0.5/assets/basics_tables.md.CJA8tShT.lean.js
@@ -1,0 +1,1 @@
+import{_ as a,c as e,o as i,aA as n}from"./chunks/framework.cOjVJm4c.js";const g=JSON.parse('{"title":"Tabular Listings of GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/tables.md","filePath":"basics/tables.md","lastUpdated":null}'),t={name:"basics/tables.md"};function o(l,s,p,r,h,d){return i(),e("div",null,s[0]||(s[0]=[n("",10)]))}const k=a(t,[["render",o]]);export{g as __pageData,k as default};
