+++ v8.0.5/assets/tutorials_using_isequal.md.BrkWE7wD.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as e,o as a,aA as n}from"./chunks/framework.cOjVJm4c.js";const g=JSON.parse('{"title":"Equivalence between GeoRegions and their Properties","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/isequal.md","filePath":"tutorials/using/isequal.md","lastUpdated":null}'),t={name:"tutorials/using/isequal.md"};function h(l,s,p,o,k,d){return a(),e("div",null,s[0]||(s[0]=[n("",44)]))}const c=i(t,[["render",h]]);export{g as __pageData,c as default};
