+++ v8.0.5/assets/api_derotation.md.Ctd_p2PX.js
@@ -1,0 +1,24 @@
+import{_ as l,C as r,c as p,o as h,j as s,a,aA as n,G as t,w as o}from"./chunks/framework.cOjVJm4c.js";const A=JSON.parse('{"title":"API for Derotations","description":"","frontmatter":{},"headers":[],"relativePath":"api/derotation.md","filePath":"api/derotation.md","lastUpdated":null}'),d={name:"api/derotation.md"},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""};function E(u,i,y,f,F,b){const e=r("Badge");return h(),p("div",null,[i[16]||(i[16]=s("h1",{id:"API-for-Derotations",tabindex:"-1"},[a("API for Derotations "),s("a",{class:"header-anchor",href:"#API-for-Derotations","aria-label":'Permalink to "API for Derotations {#API-for-Derotations}"'},"​")],-1)),i[17]||(i[17]=s("h2",{id:"Derotating-a-Shape",tabindex:"-1"},[a("Derotating a Shape "),s("a",{class:"header-anchor",href:"#Derotating-a-Shape","aria-label":'Permalink to "Derotating a Shape {#Derotating-a-Shape}"'},"​")],-1)),s("details",k,[s("summary",null,[i[0]||(i[0]=s("a",{id:"GeoRegions.derotatecoordinates",href:"#GeoRegions.derotatecoordinates"},[s("span",{class:"jlbinding"},"GeoRegions.derotatecoordinates")],-1)),i[1]||(i[1]=a()),t(e,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),i[4]||(i[4]=n(`<div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">derotatecoordinates</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    lon </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Vector{&lt;:Real}</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">,</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    lat </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Vector{&lt;:Real}</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">,</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    geo </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> GeoRegion</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">;</span></span>
[…]
