+++ v8.0.5/assets/api_tables.md.BvoBiJs9.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,C as n,c as p,o as h,j as i,a,aA as o,G as l,w as k}from"./chunks/framework.cOjVJm4c.js";const C=JSON.parse('{"title":"API for Tables in GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"api/tables.md","filePath":"api/tables.md","lastUpdated":null}'),d={name:"api/tables.md"},r={class:"jldocstring custom-block",open:""};function g(c,s,E,y,F,u){const e=n("Badge");return h(),p("div",null,[s[4]||(s[4]=i("h1",{id:"API-for-Tables-in-GeoRegions.jl",tabindex:"-1"},[a("API for Tables in GeoRegions.jl "),i("a",{class:"header-anchor",href:"#API-for-Tables-in-GeoRegions.jl","aria-label":'Permalink to "API for Tables in GeoRegions.jl {#API-for-Tables-in-GeoRegions.jl}"'},"​")],-1)),i("details",r,[i("summary",null,[s[0]||(s[0]=i("a",{id:"GeoRegions.tableGeoRegions-Tuple{}",href:"#GeoRegions.tableGeoRegions-Tuple{}"},[i("span",{class:"jlbinding"},"GeoRegions.tableGeoRegions")],-1)),s[1]||(s[1]=a()),l(e,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[3]||(s[3]=o("",4)),l(e,{type:"info",class:"source-link",text:"source"},{default:k(()=>s[2]||(s[2]=[i("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/3a61e137e6dc791264adbcdab55f23afe7ee03bd/src/georegions/tables.jl#L1-L27",target:"_blank",rel:"noreferrer"},"source",-1)])),_:1})])])}const b=t(d,[["render",g]]);export{C as __pageData,b as default};
