### 2025-03-18 #7829 @kthyng
In addition to the `cloc` above then we need to add two more for the other two repositories:

```
>>> cloc RegionGrids.jl-main.zip
      30 text files.
      28 unique files.
       6 files ignored.

Language                     files          blank        comment           code
Julia                           10            136            199            454
Markdown                         8            136              3            390
CSS                              1             55             72            157
YAML                             5              0              5            149
Vuejs Component                  2             23              1            148
TOML                             2              4              0             23
SUM:                            28            354            280           1321
```

```
>>> cloc LandSea.jl-main.zip
      20 text files.
      18 unique files.
       5 files ignored.

Language                     files          blank        comment           code
CSS                              1             55             72            157
YAML                             5              0              5            149
Vuejs Component                  2             23              1            148
Markdown                         4             22              0             94
Julia                            4             23             28             85
TOML                             2              4              1             19
SUM:                            18            127            107            652

```
