### 2025-06-25 #7829 @hugoledoux
@editorial add @mfsch as reviewer
