### 2025-07-13 3168728
Updating for the newest version of DocumenterVitepress
