+++ docs/make.jl
@@ -68,7 +68,7 @@ for file in files
     rm(file)
 end
 
-deploydocs(;
+DocumenterVitepress.deploydocs(;
     repo      = "github.com/GeoRegionsEcosystem/GeoRegions.jl.git",
     target    = "build", # this is where Vitepress stores its output
     devbranch = "main",
