+++ docs/src/api/derotation.md
@@ -1,14 +1,0 @@
-# API for Derotations
-
-## Derotating a Shape
-
-```@docs
-derotatecoordinates
-derotatecoordinates!
-```
-
-## Derotating a Point Coordinate
-
-```@docs
-derotatepoint
-```
