+++ docs/src/api/project.md
@@ -1,14 +1,0 @@
-# API for Project-Handling in GeoRegions.jl
-
-## Setting up GeoRegions.jl in a Project
-```@docs
-setupGeoRegions
-deleteGeoRegions
-```
-
-## Handling custom lists of GeoRegions.jl
-
-```@docs
-GeoRegions.loadGeoRegions
-addGeoRegions
-```

[…]
