+++ dev/assets/tutorials_using_isequal.md.EcZwDtVx.lean.js
@@ -1,1 +1,0 @@
-import{_ as i,c as e,o as a,aA as n}from"./chunks/framework.D3LA_rsn.js";const g=JSON.parse('{"title":"Equivalence between GeoRegions and their Properties","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/isequal.md","filePath":"tutorials/using/isequal.md","lastUpdated":null}'),t={name:"tutorials/using/isequal.md"};function h(l,s,p,o,k,d){return a(),e("div",null,s[0]||(s[0]=[n("",44)]))}const c=i(t,[["render",h]]);export{g as __pageData,c as default};
