+++ dev/assets/tutorials_projects_setup.md.TMFUnv7E.lean.js
@@ -1,1 +1,0 @@
-import{_ as s,c as i,o as a,aA as t}from"./chunks/framework.D3LA_rsn.js";const g=JSON.parse('{"title":"GeoRegions.jl for your Research Project","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/projects/setup.md","filePath":"tutorials/projects/setup.md","lastUpdated":null}'),o={name:"tutorials/projects/setup.md"};function n(l,e,p,h,r,d){return a(),i("div",null,e[0]||(e[0]=[t("",34)]))}const k=s(o,[["render",n]]);export{g as __pageData,k as default};
