+++ dev/assets/index.md.B71m_nU_.lean.js
@@ -1,1 +1,0 @@
-import{_ as i,c as s,o as t,aA as a}from"./chunks/framework.D3LA_rsn.js";const k=JSON.parse(`{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"GeoRegions.jl","text":"Geographic Regions for Geospatial Data","tagline":"Defining Geographic Regions of Interest for Geospatial Data in Geoscience and Climate.","image":{"src":"/logo.png","alt":"GeoRegions"},"actions":[{"theme":"brand","text":"Getting Started","link":"/georegions"},{"theme":"alt","text":"Basics","link":"/basics"},{"theme":"alt","text":"Tutorials","link":"/tutorials"},{"theme":"alt","text":"View on Github","link":"https://github.com/GeoRegionsEcosystem/GeoRegions.jl"}]},"features":[{"title":"💥 Simple","details":"Geographic Regions are defined using Longitude/Latitude coordinates.","link":"/basics/shape."},{"title":"🌍 Predefined GeoRegions","details":"Many predefined Geographic Regions bundled from different datasets included!","link":"/basics/predefined/datasets"},{"title":"⚙️ Customizable","details":"Define your own custom Geographic Regions for immediate use.","link":"/basics/create"},{"title":"🔍 Useful","details":"Check if points or a region are in/on a GeoRegion? Sure thing.","link":"/tutorials/using/isin"},{"title":"🔄 Rotation","details":"Want to \\"de\\"rotate your data for better analysis? We've got you covered!","link":"/tutorials/using/derotate"},{"title":"🗂️ Save for Later","details":"Want to save your custom GeoRegions for later? Start here!","link":"/tutorials/projects/addreadrm"}]},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}`),n={name:"index.md"};function o(l,e,h,r,p,d){return t(),s("div",null,e[0]||(e[0]=[a("",9)]))}const c=i(n,[["render",o]]);export{k as __pageData,c as default};

[…]
