+++ dev/assets/basics_read.md.BP6guYBi.lean.js
@@ -1,0 +1,1 @@
+import{_ as s,c as i,o as a,aA as n}from"./chunks/framework.B0_WgCZj.js";const g=JSON.parse('{"title":"Reading Predefined GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/read.md","filePath":"basics/read.md","lastUpdated":null}'),t={name:"basics/read.md"};function o(l,e,p,d,h,r){return a(),i("div",null,e[0]||(e[0]=[n("",9)]))}const k=s(t,[["render",o]]);export{g as __pageData,k as default};
