+++ dev/assets/api_tables.md.CvrkAO1E.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,C as n,c as p,o as h,j as i,a,aA as o,G as l,w as d}from"./chunks/framework.B0_WgCZj.js";const C=JSON.parse('{"title":"API for Tables in GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"api/tables.md","filePath":"api/tables.md","lastUpdated":null}'),k={name:"api/tables.md"},r={class:"jldocstring custom-block",open:""};function g(c,s,E,y,F,u){const e=n("Badge");return h(),p("div",null,[s[4]||(s[4]=i("h1",{id:"API-for-Tables-in-GeoRegions.jl",tabindex:"-1"},[a("API for Tables in GeoRegions.jl "),i("a",{class:"header-anchor",href:"#API-for-Tables-in-GeoRegions.jl","aria-label":'Permalink to "API for Tables in GeoRegions.jl {#API-for-Tables-in-GeoRegions.jl}"'},"​")],-1)),i("details",r,[i("summary",null,[s[0]||(s[0]=i("a",{id:"GeoRegions.tableGeoRegions-Tuple{}",href:"#GeoRegions.tableGeoRegions-Tuple{}"},[i("span",{class:"jlbinding"},"GeoRegions.tableGeoRegions")],-1)),s[1]||(s[1]=a()),l(e,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[3]||(s[3]=o("",4)),l(e,{type:"info",class:"source-link",text:"source"},{default:d(()=>s[2]||(s[2]=[i("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/de91f256366756a045ddd3fa626b69d7cdb92589/src/georegions/tables.jl#L1-L27",target:"_blank",rel:"noreferrer"},"source",-1)])),_:1,__:[2]})])])}const b=t(k,[["render",g]]);export{C as __pageData,b as default};
