+++ dev/assets/api_shape.md.0EJSUc54.js
@@ -1,0 +1,6 @@
+import{_ as n,C as o,c as r,o as l,j as i,a as s,aA as p,G as a,w as h}from"./chunks/framework.B0_WgCZj.js";const R=JSON.parse('{"title":"API for Different GeoRegion Properties","description":"","frontmatter":{},"headers":[],"relativePath":"api/shape.md","filePath":"api/shape.md","lastUpdated":null}'),d={name:"api/shape.md"},k={class:"jldocstring custom-block",open:""};function g(c,e,f,u,E,y){const t=o("Badge");return l(),r("div",null,[e[4]||(e[4]=i("h1",{id:"API-for-Different-GeoRegion-Properties",tabindex:"-1"},[s("API for Different GeoRegion Properties "),i("a",{class:"header-anchor",href:"#API-for-Different-GeoRegion-Properties","aria-label":'Permalink to "API for Different GeoRegion Properties {#API-for-Different-GeoRegion-Properties}"'},"​")],-1)),e[5]||(e[5]=i("h2",{id:"Retrieve-the-GeoRegion-Shape",tabindex:"-1"},[s("Retrieve the GeoRegion Shape "),i("a",{class:"header-anchor",href:"#Retrieve-the-GeoRegion-Shape","aria-label":'Permalink to "Retrieve the GeoRegion Shape {#Retrieve-the-GeoRegion-Shape}"'},"​")],-1)),i("details",k,[i("summary",null,[e[0]||(e[0]=i("a",{id:"GeoRegions.coordinates",href:"#GeoRegions.coordinates"},[i("span",{class:"jlbinding"},"GeoRegions.coordinates")],-1)),e[1]||(e[1]=s()),a(t,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),e[3]||(e[3]=p(`<div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">coordinates</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    geo </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> GeoRegion</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">;</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    n   </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Int</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;"> =</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> 1</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">,</span></span>
+<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">    derotate </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">::</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> Bool</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;"> =</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> false</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">,</span></span>
[…]
