+++ docs/src/api/shape.md
@@ -1,0 +1,7 @@
+# API for Different GeoRegion Properties
+
+## Retrieve the GeoRegion Shape
+
+```@docs
+coordinates
+```
