+++ docs/src/api/derotation.md
@@ -1,0 +1,14 @@
+# API for Derotations
+
+## Derotating a Shape
+
+```@docs
+derotatecoordinates
+derotatecoordinates!
+```
+
+## Derotating a Point Coordinate
+
+```@docs
+derotatepoint
+```

[…]
