+++ dev/assets/tutorials_projects_addreadrm.md.Jfoy7U3R.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as e,aA as n}from"./chunks/framework.B0_WgCZj.js";const t="/GeoRegions.jl/dev/assets/kzvgclu.sqZmjjTV.png",E=JSON.parse('{"title":"Manipulation of User-Defined GeoRegions.jl for your Project","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/projects/addreadrm.md","filePath":"tutorials/projects/addreadrm.md","lastUpdated":null}'),h={name:"tutorials/projects/addreadrm.md"};function l(k,s,p,o,d,r){return e(),a("div",null,s[0]||(s[0]=[n("",78)]))}const y=i(h,[["render",l]]);export{E as __pageData,y as default};
