### 2025-07-23 #8480 @hugoledoux
@mfsch Kind reminder about this review, and if you could generate your checklist that would be a great start
