### 2025-08-28 natgeo-wong/GeoRegions.jl#16 @boriskaus
Contributor guideline lacking
As part of the [JOSS](https://github.com/openjournals/joss-reviews/issues/8480#issuecomment-3148799447) review, I was looking through the documentation. You don't seem to have guidelines for new users that wish to contribute to the package (which the JOSS review asks for). Can this be added?
