### 2025-08-29 #8480 @hugoledoux
@natgeo-wong just a reminder that the reviewers brought feedback that is relevant, and the idea is that you answer to the questions/critics here in the issue. We want to foster a discussion that happens here.
