### 2025-09-27 5b039db
Debug the defining of GeoRegions when pID =/= GLB, bump to v8.0.6
