+++ Project.toml
@@ -1,7 +1,7 @@
 name = "GeoRegions"
 uuid = "b001f823-fa75-4bff-bf55-6610c8f3688a"
 authors = ["natgeo-wong <natgeo.wong@outlook.com>"]
-version = "8.0.5"
+version = "8.0.6"
 
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
