+++ dev/assets/tutorials_using_ison.md.CIIM5Cy2.lean.js
@@ -1,1 +1,0 @@
-import{_ as i,c as a,o as e,aA as n}from"./chunks/framework.BqlRVcpN.js";const t="/GeoRegions.jl/dev/assets/ksmqszd.C-zWTFrp.png",E=JSON.parse('{"title":"Is it on a GeoRegion Boundary?","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/ison.md","filePath":"tutorials/using/ison.md","lastUpdated":null}'),h={name:"tutorials/using/ison.md"};function l(p,s,k,o,d,r){return e(),a("div",null,[...s[0]||(s[0]=[n("",48)])])}const c=i(h,[["render",l]]);export{E as __pageData,c as default};
