+++ dev/assets/chunks/@localSearchIndexroot.V7yS8iGj.js
@@ -1,0 +1,1 @@
+const e=`{"documentCount":92,"nextId":92,"documentIds":{"0":"/GeoRegions.jl/dev/api/createaddreadrm#API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions","1":"/GeoRegions.jl/dev/api/createaddreadrm#Creating-GeoRegions","2":"/GeoRegions.jl/dev/api/createaddreadrm#Retrieving-GeoRegions","3":"/GeoRegions.jl/dev/api/createaddreadrm#Adding-Custom-GeoRegions","4":"/GeoRegions.jl/dev/api/createaddreadrm#Overwriting-Custom-GeoRegions","5":"/GeoRegions.jl/dev/api/createaddreadrm#Removing-Custom-GeoRegions","6":"/GeoRegions.jl/dev/api/createaddreadrm#After-having-loaded-a-Custom-GeoRegion","7":"/GeoRegions.jl/dev/api/createaddreadrm#Based-on-the-ID-of-a-Custom-GeoRegion","8":"/GeoRegions.jl/dev/api/derotation#API-for-Derotations","9":"/GeoRegions.jl/dev/api/derotation#Derotating-a-Shape","10":"/GeoRegions.jl/dev/api/derotation#Derotating-a-Point-Coordinate","11":"/GeoRegions.jl/dev/api/isinonequal#API-for-Is-it-in/on/equal-to-a-GeoRegion?","12":"/GeoRegions.jl/dev/api/isinonequal#In","13":"/GeoRegions.jl/dev/api/isinonequal#Is-Point-In-a-GeoRegion","14":"/GeoRegions.jl/dev/api/isinonequal#Is-GeoRegion-In-a-GeoRegion","15":"/GeoRegions.jl/dev/api/isinonequal#On","16":"/GeoRegions.jl/dev/api/isinonequal#Is-Point-On-a-GeoRegion","17":"/GeoRegions.jl/dev/api/isinonequal#Is-GeoRegion-On-a-GeoRegion","18":"/GeoRegions.jl/dev/api/isinonequal#Equivalence-of-GeoRegions","19":"/GeoRegions.jl/dev/api/isinonequal#Does-this-GeoRegion-already-Exist?","20":"/GeoRegions.jl/dev/api/isinonequal#Equivalence-of-GeoRegion-IDs","21":"/GeoRegions.jl/dev/api/isinonequal#Equivalence-of-GeoRegion-Shapes","22":"/GeoRegions.jl/dev/api/project#API-for-Project-Handling-in-GeoRegions.jl","23":"/GeoRegions.jl/dev/api/project#Setting-up-GeoRegions.jl-in-a-Project","24":"/GeoRegions.jl/dev/api/project#Handling-custom-lists-of-GeoRegions.jl","25":"/GeoRegions.jl/dev/api/shape#API-for-Different-GeoRegion-Properties","26":"/GeoRegions.jl/dev/api/shape#Retrieve-the-GeoRegion-Shape","27":"/GeoRegions.jl/dev/basics/create#Creating-Custom-GeoRegions","28":"/GeoRegions.jl/dev/api/tables#API-for-Tables-in-GeoRegions.jl","29":"/GeoRegions.jl/dev/basics/#the-basics","30":"/GeoRegions.jl/dev/basics/#predefined-georegions","31":"/GeoRegions.jl/dev/basics/predefined/datasets#Predefined-GeoRegions-in-GeoRegions.jl","32":"/GeoRegions.jl/dev/basics/predefined/datasets#Adapted-from-Giorgi-and-Francisco-2000","33":"/GeoRegions.jl/dev/basics/predefined/datasets#SREX-Regions-adapted-from-Seneviratne-
[…]
