### 2025-10-12 90920e1
Updating the PrettyTables [compat]
