+++ dev/assets/georegions.md.bhhM5ZSz.lean.js
@@ -1,1 +1,0 @@
-import{_ as a,C as r,c as d,o as c,aA as s,j as o,G as n,a as t,w as l}from"./chunks/framework.BqlRVcpN.js";const _=JSON.parse('{"title":"What is a GeoRegion?","description":"","frontmatter":{},"headers":[],"relativePath":"georegions.md","filePath":"georegions.md","lastUpdated":null}'),p={name:"georegions.md"},g={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},h={class:"jldocstring custom-block",open:""};function y(G,e,f,m,b,T){const i=r("Badge");return c(),d("div",null,[e[12]||(e[12]=s("",4)),o("details",g,[o("summary",null,[e[0]||(e[0]=o("a",{id:"GeoRegions.AbstractGeoRegion",href:"#GeoRegions.AbstractGeoRegion"},[o("span",{class:"jlbinding"},"GeoRegions.AbstractGeoRegion")],-1)),e[1]||(e[1]=t()),n(i,{type:"info",class:"jlObjectType jlType",text:"Type"})]),e[3]||(e[3]=s("",3)),n(i,{type:"info",class:"source-link",text:"source"},{default:l(()=>[...e[2]||(e[2]=[o("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/5b039db4afc150c0c3e68753b8ef289e3310d2c1/src/GeoRegions.jl#L28-L42",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),e[13]||(e[13]=s("",6)),o("details",u,[o("summary",null,[e[4]||(e[4]=o("a",{id:"GeoRegions.GeoRegion",href:"#GeoRegions.GeoRegion"},[o("span",{class:"jlbinding"},"GeoRegions.GeoRegion")],-1)),e[5]||(e[5]=t()),n(i,{type:"info",class:"jlObjectType jlType",text:"Type"})]),e[7]||(e[7]=s("",3)),n(i,{type:"info",class:"source-link",text:"source"},{default:l(()=>[...e[6]||(e[6]=[o("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/5b039db4afc150c0c3e68753b8ef289e3310d2c1/src/GeoRegions.jl#L85-L90",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),e[14]||(e[14]=o("h2",{id:"The-Geometry-Type",tabindex:"-1"},[t("The "),o("code",null,"Geometry"),t(" Type "),o("a",{class:"header-anchor",href:"#The-Geometry-Type","aria-label":'Permalink to "The `Geometry` Type {#The-Geometry-Type}"'},"​")],-1)),e[15]||(e[15]=o("p",null,[t("All "),o("code",null,"GeoRegion"),t("s are defined based on a "),o("code",null,"Geometry"),t(" struct type that contains information on the geometry of a GeoRegion.")],-1)),o("details",h,[o("summary",null,[e[8]||(e[8]=o("a",{id:"GeoRegions.Geometry",href:"#GeoRegions.Geometry"},[o("span",{class:"jlbinding"},"GeoRegions.Geometry")],-1)),e[9]||(e[9]=t()),n(i,{type:"info",class:"jlObjectType jlType",text:"Type"})]),e[11]||(e[11]=s("",3)),n(i,{type:"info",class:"source-link",text:"source"},{default:l(()=>[...e[10]||(e[10]=[o("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/5b039db4afc150c0c3e68753b8ef289e3310d2c1/src/GeoRegions.jl#
[…]
