+++ dev/assets/api_shape.md.BIY_YJgS.lean.js
@@ -1,1 +1,0 @@
-import{_ as n,C as o,c as r,o as l,j as i,a as s,aA as p,G as a,w as h}from"./chunks/framework.BqlRVcpN.js";const R=JSON.parse('{"title":"API for Different GeoRegion Properties","description":"","frontmatter":{},"headers":[],"relativePath":"api/shape.md","filePath":"api/shape.md","lastUpdated":null}'),d={name:"api/shape.md"},k={class:"jldocstring custom-block",open:""};function g(c,e,f,u,E,y){const t=o("Badge");return l(),r("div",null,[e[4]||(e[4]=i("h1",{id:"API-for-Different-GeoRegion-Properties",tabindex:"-1"},[s("API for Different GeoRegion Properties "),i("a",{class:"header-anchor",href:"#API-for-Different-GeoRegion-Properties","aria-label":'Permalink to "API for Different GeoRegion Properties {#API-for-Different-GeoRegion-Properties}"'},"​")],-1)),e[5]||(e[5]=i("h2",{id:"Retrieve-the-GeoRegion-Shape",tabindex:"-1"},[s("Retrieve the GeoRegion Shape "),i("a",{class:"header-anchor",href:"#Retrieve-the-GeoRegion-Shape","aria-label":'Permalink to "Retrieve the GeoRegion Shape {#Retrieve-the-GeoRegion-Shape}"'},"​")],-1)),i("details",k,[i("summary",null,[e[0]||(e[0]=i("a",{id:"GeoRegions.coordinates",href:"#GeoRegions.coordinates"},[i("span",{class:"jlbinding"},"GeoRegions.coordinates")],-1)),e[1]||(e[1]=s()),a(t,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),e[3]||(e[3]=p("",8)),a(t,{type:"info",class:"source-link",text:"source"},{default:h(()=>[...e[2]||(e[2]=[i("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/5b039db4afc150c0c3e68753b8ef289e3310d2c1/src/georegions/shape.jl#L1-L25",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})])])}const m=n(d,[["render",g]]);export{R as __pageData,m as default};

[…]
