+++ dev/assets/georegions.md.CwRcMZ0E.js
@@ -1,1 +1,0 @@
-import{_ as a,C as r,c as d,o as c,aA as s,j as o,G as n,a as t,w as l}from"./chunks/framework.BqlRVcpN.js";const _=JSON.parse('{"title":"What is a GeoRegion?","description":"","frontmatter":{},"headers":[],"relativePath":"georegions.md","filePath":"georegions.md","lastUpdated":null}'),p={name:"georegions.md"},g={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},h={class:"jldocstring custom-block",open:""};function y(f,e,G,m,b,T){const i=r("Badge");return c(),d("div",null,[e[12]||(e[12]=s('<h1 id="What-is-a-GeoRegion?" tabindex="-1">What is a GeoRegion? <a class="header-anchor" href="#What-is-a-GeoRegion?" aria-label="Permalink to &quot;What is a GeoRegion? {#What-is-a-GeoRegion?}&quot;">​</a></h1><p>In essence, a <code>GeoRegion</code> is:</p><ul><li><p>a geographical region of interest</p></li><li><p>identified by an <code>ID</code></p></li><li><p>a subregion of a <strong>parent</strong> <code>GeoRegion</code> (identified by <code>pID</code>, which must also be a valid <code>ID</code>)</p></li></ul><div class="tip custom-block"><p class="custom-block-title">Default GeoRegions</p><p>When using GeoRegions.jl, the default <code>GeoRegion</code> should generally be the global domain, specified by <code>GLB</code> and given by the <code>[N,S,E,W]</code> coordinates <code>[90,-90,360,0]</code>. The Global GeoRegion <code>GLB</code> is considered to be a subset of itself.</p></div>',4)),o("details",g,[o("summary",null,[e[0]||(e[0]=o("a",{id:"GeoRegions.AbstractGeoRegion",href:"#GeoRegions.AbstractGeoRegion"},[o("span",{class:"jlbinding"},"GeoRegions.AbstractGeoRegion")],-1)),e[1]||(e[1]=t()),n(i,{type:"info",class:"jlObjectType jlType",text:"Type"})]),e[3]||(e[3]=s('<div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">AbstractGeoRegion</span></span></code></pre></div><p>Abstract supertype for geographical regions. All <code>AbstractGeoRegion</code> types contain the following fields:</p><ul><li><p><code>ID</code> - A <code>String</code> Type, the identifier for the GeoRegion.</p></li><li><p><code>pID</code> - A <code>String</code> Type, the identifier for the parent GeoRegion.</p></li><li><p><code>name</code> - A <code>String</code> Type, the full name of the GeoRegion.</p></li><li><p><code>path</code> - A <code>String</code> Type, the path of the file containing information on this GeoRegion.</p></li><li><p><code>N</code> - A <code>Float</code> Type, contains the northernmost latitude bound.</p></li><li><p><code>S</code> - A <code>Float</code> Type, contains the southernmost latitude bound.</p></li><li><p><code>E</code> - A <code>Float</code> Type, contains the easternmost longitude bound.</p></li><li><p><code>W</code> - A <code>Float</code> Type, contains the westernmost longitude bound.</p></li><li><p><code>θ</code> - A <code>Float<
[…]
