+++ stable/assets/basics_tables.md.ZSPIPZpC.js
@@ -1,0 +1,14 @@
+import{_ as i,c as e,o as a,aA as t}from"./chunks/framework.yEV9BkZI.js";const c=JSON.parse('{"title":"Tabular Listings of GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/tables.md","filePath":"basics/tables.md","lastUpdated":null}'),n={name:"basics/tables.md"};function l(o,s,h,p,r,d){return a(),e("div",null,[...s[0]||(s[0]=[t(`<h1 id="Tabular-Listings-of-GeoRegions" tabindex="-1">Tabular Listings of GeoRegions <a class="header-anchor" href="#Tabular-Listings-of-GeoRegions" aria-label="Permalink to &quot;Tabular Listings of GeoRegions {#Tabular-Listings-of-GeoRegions}&quot;">​</a></h1><p>A list of GeoRegions and their basic properties can be called using the function <code>tableGeoRegions()</code>, which lists all available GeoRegions, both predefined and <a href="/GeoRegions.jl/stable/basics/create">user-defined</a>. You can also choose to list only predefined or user-customized GeoRegions by setting their respective keywords arguments to <code>true/false</code>.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">tableGeoRegions</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(predefined </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> true</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">/</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">false</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">, custom </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> true</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">/</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">false</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span></code></pre></div><div class="tip custom-block"><p class="custom-block-title">Additional Functionality for <code>tableGeoRegions()</code></p><p>It is possible to use <code>tableGeoRegions()</code> to list user-defined <code>GeoRegion</code>s in specific paths/directories. By default, it will list user-defined <code>GeoRegion</code>s saved into the path <code>joinpath(homedir(),&quot;.georegions&quot;)</code>. For more on user-defined GeoRegions and saving them as part of a larger project, refer to how to <a href="/GeoRegions.jl/stable/tutorials/projects/setup">setup</a> GeoRegions.jl for a project.</p></div><h2 id="Use-Case-Examples" tabindex="-1">Use Case Examples <a class="header-anchor" href="#Use-Case-Examples" aria-label="Permalink to &quot;Use Case Examples {#Use-Case-Examples}&quot;">​</a></h2><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style
[…]
