+++ stable/assets/basics_index.md.DYJ_ihDC.lean.js
@@ -1,0 +1,1 @@
+import{G as s}from"./chunks/Gallery.DXMhafAj.js";import{d as o,c as d,o as c,j as a,G as i,a as n}from"./chunks/framework.yEV9BkZI.js";const m=JSON.parse('{"title":"The Basics","description":"","frontmatter":{},"headers":[],"relativePath":"basics/index.md","filePath":"basics/index.md","lastUpdated":null}'),l={name:"basics/index.md"},b=o({...l,setup(p){const t=[{href:"read",src:"images/read.png",caption:"Read / Load",desc:"How to read/load a GeoRegion into the workspace."},{href:"create",src:"images/create.png",caption:"Create",desc:"How to create a user-defined GeoRegion."},{href:"shape",src:"images/shape.png",caption:"Shape Properties",desc:"Retrieving the Shape of a GeoRegion"},{href:"tables",src:"images/list.png",caption:"Tables",desc:"Listing existing GeoRegions in Table Format"}],r=[{href:"predefined/datasets",src:"images/sets.png",caption:"Available Datasets",desc:"Different Predefined Datasets in GeoRegions.jl"},{href:"predefined/listall",src:"images/listall.png",caption:"All Predefined GeoRegions",desc:"Listing out all predefined GeoRegions in GeoRegions.jl"}];return(g,e)=>(c(),d("div",null,[e[0]||(e[0]=a("h1",{id:"the-basics",tabindex:"-1"},[n("The Basics "),a("a",{class:"header-anchor",href:"#the-basics","aria-label":'Permalink to "The Basics"'},"​")],-1)),i(s,{images:t}),e[1]||(e[1]=a("h2",{id:"predefined-georegions",tabindex:"-1"},[n("Predefined GeoRegions "),a("a",{class:"header-anchor",href:"#predefined-georegions","aria-label":'Permalink to "Predefined GeoRegions"'},"​")],-1)),i(s,{images:r})]))}});export{m as __pageData,b as default};
