+++ stable/assets/basics_tables.md.ZSPIPZpC.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as e,o as a,aA as t}from"./chunks/framework.yEV9BkZI.js";const c=JSON.parse('{"title":"Tabular Listings of GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/tables.md","filePath":"basics/tables.md","lastUpdated":null}'),n={name:"basics/tables.md"};function l(o,s,h,p,r,d){return a(),e("div",null,[...s[0]||(s[0]=[t("",10)])])}const g=i(n,[["render",l]]);export{c as __pageData,g as default};

[…]
