+++ stable/assets/ecosystem.md.5ACaOd23.js
@@ -1,1 +1,0 @@
-import{V as r,a as i,b as o,c}from"./chunks/theme.DhAyZgqz.js";import{c as p,o as u,G as t,w as s,k as n,a}from"./chunks/framework.yEV9BkZI.js";const y=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"page"},"headers":[],"relativePath":"ecosystem.md","filePath":"ecosystem.md","lastUpdated":null}'),d={name:"ecosystem.md"},G=Object.assign(d,{setup(h){const l=[{avatar:"https://github.com/GeoRegionsEcosystem.png",name:"GeoRegions.jl",desc:"Defining Geographic Regions of Interest for Geospatial Data in Geoscience and Climate.",links:[{icon:"github",link:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl"}]}],g=[{avatar:"https://github.com/GeoRegionsEcosystem.png",name:"RegionGrids.jl",desc:"Extracting Gridded Data for a given GeoRegion",links:[{icon:"github",link:"https://github.com/GeoRegionsEcosystem/RegionGrids.jl"}]},{avatar:"https://github.com/GeoRegionsEcosystem.png",name:"LandSea.jl",desc:"Extensible Land-Sea Mask Datasets",links:[{icon:"github",link:"https://github.com/GeoRegionsEcosystem/LandSea.jl"}]}],m=[{avatar:"https://raw.githubusercontent.com/natgeo-wong/NASAPrecipitation.jl/main/src/logosmall.png",name:"NASAPrecipitation.jl",desc:"Downloading, Extracting and Analyzing NASA Satellite Precipitation Data",links:[{icon:"github",link:"https://github.com/GeoRegionsEcosystem/NASAPrecipitation.jl"}]},{avatar:"https://raw.githubusercontent.com/natgeo-wong/NASAPrecipitation.jl/main/src/logosmall.png",name:"ERA5Reanalysis.jl",desc:"Downloading, Extracting and Analyzing ERA5 Reanalysis Data",links:[{icon:"github",link:"https://github.com/GeoRegionsEcosystem/ERA5Reanalysis.jl"}]},{avatar:"https://raw.githubusercontent.com/natgeo-wong/NASAPrecipitation.jl/main/src/logosmall.png",name:"NASABrightnessTemperature.jl",desc:"Downloading, Extracting and Analyzing NASA Brightness Temperature Data",links:[{icon:"github",link:"https://github.com/GeoRegionsEcosystem/NASABrightnessTemperature.jl"}]}];return(b,e)=>(u(),p("div",null,[t(n(c),null,{default:s(()=>[t(n(r),null,{title:s(()=>[...e[0]||(e[0]=[a("Ecosystem",-1)])]),_:1}),t(n(i),null,{title:s(()=>[...e[1]||(e[1]=[a("GeoRegions.jl",-1)])]),members:s(()=>[t(n(o),{size:"small",members:l})]),_:1}),t(n(i),null,{title:s(()=>[...e[2]||(e[2]=[a("Frameworks Extending GeoRegions.jl",-1)])]),members:s(()=>[t(n(o),{size:"small",members:g})]),_:1}),t(n(i),null,{title:s(()=>[...e[3]||(e[3]=[a("The GeoRegions Ecosystem",-1)])]),members:s(()=>[t(n(o),{size:"small",members:m})]),_:1})]),_:1})]))}});export{y as __pageData,G as default};
