+++ stable/assets/api_derotation.md.gkKoaNQp.lean.js
@@ -1,0 +1,1 @@
+import{_ as l,C as r,c as p,o as h,j as s,a,aA as n,G as t,w as o}from"./chunks/framework.yEV9BkZI.js";const A=JSON.parse('{"title":"API for Derotations","description":"","frontmatter":{},"headers":[],"relativePath":"api/derotation.md","filePath":"api/derotation.md","lastUpdated":null}'),d={name:"api/derotation.md"},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""};function E(u,i,y,f,F,C){const e=r("Badge");return h(),p("div",null,[i[16]||(i[16]=s("h1",{id:"API-for-Derotations",tabindex:"-1"},[a("API for Derotations "),s("a",{class:"header-anchor",href:"#API-for-Derotations","aria-label":'Permalink to "API for Derotations {#API-for-Derotations}"'},"​")],-1)),i[17]||(i[17]=s("h2",{id:"Derotating-a-Shape",tabindex:"-1"},[a("Derotating a Shape "),s("a",{class:"header-anchor",href:"#Derotating-a-Shape","aria-label":'Permalink to "Derotating a Shape {#Derotating-a-Shape}"'},"​")],-1)),s("details",k,[s("summary",null,[i[0]||(i[0]=s("a",{id:"GeoRegions.derotatecoordinates",href:"#GeoRegions.derotatecoordinates"},[s("span",{class:"jlbinding"},"GeoRegions.derotatecoordinates")],-1)),i[1]||(i[1]=a()),t(e,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),i[4]||(i[4]=n("",8)),t(e,{type:"info",class:"source-link",text:"source"},{default:o(()=>[...i[2]||(i[2]=[s("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/90920e11a1fb1d206a7217664b569a15ef735ce8/src/georegions/derotate.jl#L1-L25",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1}),i[5]||(i[5]=n("",8)),t(e,{type:"info",class:"source-link",text:"source"},{default:o(()=>[...i[3]||(i[3]=[s("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/90920e11a1fb1d206a7217664b569a15ef735ce8/src/georegions/derotate.jl#L82-L104",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),s("details",g,[s("summary",null,[i[6]||(i[6]=s("a",{id:"GeoRegions.derotatecoordinates!",href:"#GeoRegions.derotatecoordinates!"},[s("span",{class:"jlbinding"},"GeoRegions.derotatecoordinates!")],-1)),i[7]||(i[7]=a()),t(e,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),i[9]||(i[9]=n("",6)),t(e,{type:"info",class:"source-link",text:"source"},{default:o(()=>[...i[8]||(i[8]=[s("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/90920e11a1fb1d206a7217664b569a15ef735ce8/src/georegions/derotate.jl#L45-L64",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),i[18]||(i[18]=s("h2",{id:"Derotating-a-Point-Coordinate",tabindex:"-1"},[a("Derotating a Point Coordinate "),s("a",{class:"header-anchor",href:"#Derotating-a-Poin
[…]
