+++ stable/assets/basics_predefined_listall.md.BV-wlywc.js
@@ -1,27 +1,4 @@
-<!DOCTYPE html>
-<html lang="en-US" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <title>Default List of Predefined GeoRegions | GeoRegions.jl</title>
-    <meta name="description" content="Documentation for GeoRegions.jl">
-    <meta name="generator" content="VitePress v1.6.4">
-    <link rel="preload stylesheet" href="/GeoRegions.jl/stable/assets/style.DtPhMGX_.css" as="style">
-    <link rel="preload stylesheet" href="/GeoRegions.jl/stable/vp-icons.css" as="style">
-    
-    <script type="module" src="/GeoRegions.jl/stable/assets/app.BboSSkjh.js"></script>
-    <link rel="preload" href="/GeoRegions.jl/stable/assets/inter-roman-latin.Di8DUHzh.woff2" as="font" type="font/woff2" crossorigin="">
-    <link rel="modulepreload" href="/GeoRegions.jl/stable/assets/chunks/theme.CpBnR73M.js">
-    <link rel="modulepreload" href="/GeoRegions.jl/stable/assets/chunks/framework.yEV9BkZI.js">
-    <link rel="modulepreload" href="/GeoRegions.jl/stable/assets/basics_predefined_listall.md.BV-wlywc.lean.js">
-    <link rel="icon" href="/GeoRegions.jl/stable/favicon.ico">
-    <script src="/GeoRegions.jl/versions.js"></script>
-    <script src="/GeoRegions.jl/stable/siteinfo.js"></script>
-    <script id="check-dark-mode">(()=>{const e=localStorage.getItem("vitepress-theme-appearance")||"auto",a=window.matchMedia("(prefers-color-scheme: dark)").matches;(!e||e==="auto"?a:e==="dark")&&document.documentElement.classList.add("dark")})();</script>
-    <script id="check-mac-os">document.documentElement.classList.toggle("mac",/Mac|iPhone|iPod|iPad/i.test(navigator.platform));</script>
-  </head>
-  <body>
[…]
