+++ stable/assets/tutorials_using_ison.md.DVQRU0Zi.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as e,aA as n}from"./chunks/framework.yEV9BkZI.js";const t="/GeoRegions.jl/stable/assets/nekjqjw.C-zWTFrp.png",E=JSON.parse('{"title":"Is it on a GeoRegion Boundary?","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/ison.md","filePath":"tutorials/using/ison.md","lastUpdated":null}'),h={name:"tutorials/using/ison.md"};function l(p,s,k,o,d,r){return e(),a("div",null,[...s[0]||(s[0]=[n("",48)])])}const c=i(h,[["render",l]]);export{E as __pageData,c as default};
