+++ stable/assets/chunks/theme.BQc4sbKw.js
@@ -1,0 +1,2 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/chunks/VPLocalSearchBox.DK0wVPjU.js","assets/chunks/framework.yEV9BkZI.js"])))=>i.map(i=>d[i]);
+import{d as S,c as h,r as p,n as A,o as i,a as J,t as M,b as L,w as m,T as re,e as b,_ as V,u as Se,i as Et,f as Wt,g as Ue,h as P,j as c,k as s,l as le,m as X,p as w,q as W,s as Le,v as D,x as xe,y as De,z as Ht,A as Bt,F as O,B as F,C as ae,D as it,E as we,G as $,H as Y,I as lt,J as ve,K as te,L as z,M as Ot,N as ce,O as Ie,P as Pe,Q as rt,R as Ve,S as It,U as Ct,V as Rt,W as ut,X as ze,Y as jt,Z as Ft,$ as Ut,a0 as Dt,a1 as ct,a2 as de,a3 as zt,a4 as dt,a5 as Kt,a6 as Me,a7 as ht,a8 as ne,a9 as Ce,aa as pt,ab as Ke,ac as Ge,ad as qe,ae as he,af as Gt,ag as qt,ah as Jt,ai as vt,aj as Yt,ak as Xt,al as Qt,am as He}from"./framework.yEV9BkZI.js";const Zt=S({__name:"VPBadge",props:{text:{},type:{default:"tip"}},setup(e){return(t,a)=>(i(),h("span",{class:A(["VPBadge",e.type])},[p(t.$slots,"default",{},()=>[J(M(e.text),1)])],2))}}),ea={key:0,class:"VPBackdrop"},ta=S({__name:"VPBackdrop",props:{show:{type:Boolean}},setup(e){return(t,a)=>(i(),L(re,{name:"fade"},{default:m(()=>[e.show?(i(),h("div",ea)):b("",!0)]),_:1}))}}),aa=V(ta,[["__scopeId","data-v-b06cdb19"]]),N=Se;function na(e,t){let a,o=!1;return()=>{a&&clearTimeout(a),o?a=setTimeout(e,t):(e(),(o=!0)&&setTimeout(()=>o=!1,t))}}function Re(e){return e.startsWith("/")?e:`/${e}`}function Je(e){const{pathname:t,search:a,hash:o,protocol:n}=new URL(e,"http://a.com");if(Et(e)||e.startsWith("#")||!n.startsWith("http")||!Wt(t))return e;const{site:u}=N(),l=t.endsWith("/")||t.endsWith(".html")?e:e.replace(/(?:(^\.+)\/)?.*$/,`$1${t.replace(/(\.md)?$/,u.value.cleanUrls?"":".html")}${a}${o}`);return Ue(l)}function be({correspondingLink:e=!1}={}){const{site:t,localeIndex:a,page:o,theme:n,hash:u}=N(),l=P(()=>{var r,v;return{label:(r=t.value.locales[a.value])==null?void 0:r.label,link:((v=t.value.locales[a.value])==null?void 0:v.link)||(a.value==="root"?"/":`/${a.value}/`)}});return{localeLinks:P(()=>Object.entries(t.value.locales).flatMap(([r,v])=>l.value.label===v.label?[]:{text:v.label,link:oa(v.link||(r==="root"?"/":`/${r}/`),n.value.i18nRouting!==!1&&e,o.value.relativePath.slice(l.value.link.length-1),!t.value.cleanUrls)+u.value})),currentLang:l}}function oa(e,t,a,o){return t?e.replace(/\/$/,"")+Re(a.replace(/(^|\/)index\.md$/,"$1").replace(/\.md$/,o?
[…]
