+++ stable/assets/tutorials_using_isin.md.BVzEBs64.lean.js
@@ -1,1 +1,0 @@
-import{_ as i,c as a,o as h,aA as n}from"./chunks/framework.yEV9BkZI.js";const k="/GeoRegions.jl/stable/assets/nzfpgfw.VBndSf0w.png",t="/GeoRegions.jl/stable/assets/lxwfmyx.CGSfBt3k.png",o=JSON.parse('{"title":"Is it in a GeoRegion?","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/using/isin.md","filePath":"tutorials/using/isin.md","lastUpdated":null}'),l={name:"tutorials/using/isin.md"};function p(e,s,E,d,g,r){return h(),a("div",null,[...s[0]||(s[0]=[n("",34)])])}const F=i(l,[["render",p]]);export{o as __pageData,F as default};
