+++ stable/assets/basics_shape.md.DM7cVjxd.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as e,aA as n}from"./chunks/framework.yEV9BkZI.js";const t="/GeoRegions.jl/stable/assets/gdwjjpl.Bc_XcPih.png",E=JSON.parse('{"title":"The Shape of a GeoRegion","description":"","frontmatter":{},"headers":[],"relativePath":"basics/shape.md","filePath":"basics/shape.md","lastUpdated":null}'),h={name:"basics/shape.md"};function l(p,s,k,o,d,r){return e(),a("div",null,[...s[0]||(s[0]=[n("",28)])])}const c=i(h,[["render",l]]);export{E as __pageData,c as default};

[…]
