+++ v8/assets/api_createaddreadrm.md.DthshlBj.lean.js
@@ -1,0 +1,1 @@
+import{_ as l,C as h,c as p,o as r,j as i,a as e,aA as t,G as n,w as o}from"./chunks/framework.CaKmHBHp.js";const G=JSON.parse('{"title":"API for Creating, Reading, Adding and Removing GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"api/createaddreadrm.md","filePath":"api/createaddreadrm.md","lastUpdated":null}'),k={name:"api/createaddreadrm.md"},d={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""};function m(F,s,C,f,b,R){const a=h("Badge");return r(),p("div",null,[s[24]||(s[24]=i("h1",{id:"API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions",tabindex:"-1"},[e("API for Creating, Reading, Adding and Removing GeoRegions "),i("a",{class:"header-anchor",href:"#API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions","aria-label":'Permalink to "API for Creating, Reading, Adding and Removing GeoRegions {#API-for-Creating,-Reading,-Adding-and-Removing-GeoRegions}"'},"​")],-1)),s[25]||(s[25]=i("h2",{id:"Creating-GeoRegions",tabindex:"-1"},[e("Creating GeoRegions "),i("a",{class:"header-anchor",href:"#Creating-GeoRegions","aria-label":'Permalink to "Creating GeoRegions {#Creating-GeoRegions}"'},"​")],-1)),i("details",d,[i("summary",null,[s[0]||(s[0]=i("a",{id:"GeoRegions.GeoRegion-Tuple{Vector{<:Real}, Vector{<:Real}}",href:"#GeoRegions.GeoRegion-Tuple{Vector{<:Real}, Vector{<:Real}}"},[i("span",{class:"jlbinding"},"GeoRegions.GeoRegion")],-1)),s[1]||(s[1]=e()),n(a,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[3]||(s[3]=t("",8)),n(a,{type:"info",class:"source-link",text:"source"},{default:o(()=>[...s[2]||(s[2]=[i("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/90920e11a1fb1d206a7217664b569a15ef735ce8/src/georegions/define.jl#L54-L95",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),s[26]||(s[26]=i("h2",{id:"Retrieving-GeoRegions",tabindex:"-1"},[e("Retrieving GeoRegions "),i("a",{class:"header-anchor",href:"#Retrieving-GeoRegions","aria-label":'Permalink to "Retrieving GeoRegions {#Retrieving-GeoRegions}"'},"​")],-1)),i("details",g,[i("summary",null,[s[4]||(s[4]=i("a",{id:"GeoRegions.GeoRegion-Tuple{AbstractString}",href:"#GeoRegions.GeoRegion-Tuple{AbstractString}"},[i("span",{class:"jlbinding"},"GeoRegions.GeoRegion")],-1)),s[5]||(s[5]=e()),n(a,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[7]||(s[7]=t("",8)),n(a,{type:"info",class:"source-lin
[…]
