+++ v8/assets/basics_predefined_datasets.md.C4OrRcz6.js
@@ -1,0 +1,28 @@
+import{_ as s,c as i,o as a,aA as n}from"./chunks/framework.CaKmHBHp.js";const k=JSON.parse('{"title":"Predefined GeoRegions in GeoRegions.jl","description":"","frontmatter":{},"headers":[],"relativePath":"basics/predefined/datasets.md","filePath":"basics/predefined/datasets.md","lastUpdated":null}'),t={name:"basics/predefined/datasets.md"};function l(o,e,p,d,h,r){return a(),i("div",null,[...e[0]||(e[0]=[n(`<h1 id="Predefined-GeoRegions-in-GeoRegions.jl" tabindex="-1">Predefined <code>GeoRegion</code>s in GeoRegions.jl <a class="header-anchor" href="#Predefined-GeoRegions-in-GeoRegions.jl" aria-label="Permalink to &quot;Predefined \`GeoRegion\`s in GeoRegions.jl {#Predefined-GeoRegions-in-GeoRegions.jl}&quot;">​</a></h1><p>In GeoRegions.jl, there are several predefined sets of <code>GeoRegion</code>s.</p><h2 id="Adapted-from-Giorgi-and-Francisco-2000" tabindex="-1">Adapted from Giorgi &amp; Francisco [2000] <a class="header-anchor" href="#Adapted-from-Giorgi-and-Francisco-2000" aria-label="Permalink to &quot;Adapted from Giorgi &amp;amp; Francisco [2000] {#Adapted-from-Giorgi-and-Francisco-2000}&quot;">​</a></h2><p>Add the <code>GF_</code> prefix to the 3-letter IDs given in Giorgi &amp; Francisco [2000] to call GeoRegions adapted from this paper.</p><div class="tip custom-block"><p class="custom-block-title">Defined in GeoRegions.jl ≥ v1.1</p><p>All <code>GF_*</code> domains are defined only in versions ≥1.1</p></div><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
+<span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">GeoRegion</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;">&quot;GF_WAF&quot;</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span></code></pre></div><div class="language- vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang"></span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#24292e;--shiki-dark:#e1e4e8;">The GeoRegion GF_WAF has the following properties:</span></span>
+<span class="line"><span style="--shiki-light:#24292e;--shiki-dark:#e1e4e8;">    Region ID     (ID) : GF_WAF</span></span>
+<span class="line"><span style="--shiki-light:#24292e;--shiki-dark:#e1e4e8;">    Parent ID    (pID) : GLB</span></span>
+<span class="line"><span style="--shiki-light:#24292e;--shiki-dark:#e1e4e8;">    Name        (name) : Western Africa</span></span>
[…]
