+++ v8/assets/basics_create.md.BrYzqmcm.lean.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,o as e,aA as n}from"./chunks/framework.CaKmHBHp.js";const g=JSON.parse('{"title":"Creating Custom GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/create.md","filePath":"basics/create.md","lastUpdated":null}'),t={name:"basics/create.md"};function h(l,s,p,k,r,o){return e(),a("div",null,[...s[0]||(s[0]=[n("",5)])])}const c=i(t,[["render",h]]);export{g as __pageData,c as default};

[…]
