+++ v8/assets/basics_predefined_listall.md.BW6BZc8I.js
@@ -1,131 +1,0 @@
-import{_ as i,c as e,o as a,aA as n}from"./chunks/framework.CaKmHBHp.js";const g=JSON.parse('{"title":"Default List of Predefined GeoRegions","description":"","frontmatter":{},"headers":[],"relativePath":"basics/predefined/listall.md","filePath":"basics/predefined/listall.md","lastUpdated":null}'),l={name:"basics/predefined/listall.md"};function t(h,s,p,k,r,d){return a(),e("div",null,[...s[0]||(s[0]=[n(`<h1 id="Default-List-of-Predefined-GeoRegions" tabindex="-1">Default List of Predefined GeoRegions <a class="header-anchor" href="#Default-List-of-Predefined-GeoRegions" aria-label="Permalink to &quot;Default List of Predefined GeoRegions {#Default-List-of-Predefined-GeoRegions}&quot;">​</a></h1><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> GeoRegions</span></span>
-<span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">setupGeoRegions</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(overwrite</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">true</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">);</span></span></code></pre></div><div class="language- vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang"></span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#dbab09;--shiki-light-font-weight:bold;--shiki-dark:#ffea7f;--shiki-dark-font-weight:bold;">┌ </span><span style="--shiki-light:#dbab09;--shiki-light-font-weight:bold;--shiki-dark:#ffea7f;--shiki-dark-font-weight:bold;">Warning: </span><span style="--shiki-light:#24292e;--shiki-dark:#e1e4e8;">2025-10-12T23:05:14.211 - GeoRegions.jl - Removing custom GeoRegions.jl files from /home/runner/work/GeoRegions.jl/GeoRegions.jl/docs/build/basics/predefined/.georegions, all GeoRegion information saved into these files will be permanently lost.</span></span>
[…]
