+++ v8/assets/api_isinonequal.md.DVLxRbv-.lean.js
@@ -1,0 +1,1 @@
+import{_ as l,C as h,c as p,o as d,aA as t,j as i,G as a,a as n,w as o}from"./chunks/framework.CaKmHBHp.js";const T=JSON.parse('{"title":"API for Is it in/on/equal to a GeoRegion?","description":"","frontmatter":{},"headers":[],"relativePath":"api/isinonequal.md","filePath":"api/isinonequal.md","lastUpdated":null}'),r={name:"api/isinonequal.md"},k={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},f={class:"jldocstring custom-block",open:""},F={class:"jldocstring custom-block",open:""},b={class:"jldocstring custom-block",open:""},C={class:"jldocstring custom-block",open:""};function R(G,s,A,m,D,B){const e=h("Badge");return d(),p("div",null,[s[40]||(s[40]=t("",3)),i("details",k,[i("summary",null,[s[0]||(s[0]=i("a",{id:"Base.in-Tuple{Point2{<:Real}, GeoRegion}",href:"#Base.in-Tuple{Point2{<:Real}, GeoRegion}"},[i("span",{class:"jlbinding"},"Base.in")],-1)),s[1]||(s[1]=n()),a(e,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[3]||(s[3]=t("",8)),a(e,{type:"info",class:"source-link",text:"source"},{default:o(()=>[...s[2]||(s[2]=[i("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/5b039db4afc150c0c3e68753b8ef289e3310d2c1/src/is/isin.jl#L1-L24",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),s[41]||(s[41]=i("h3",{id:"Is-GeoRegion-In-a-GeoRegion",tabindex:"-1"},[n("Is GeoRegion In a GeoRegion "),i("a",{class:"header-anchor",href:"#Is-GeoRegion-In-a-GeoRegion","aria-label":'Permalink to "Is GeoRegion In a GeoRegion {#Is-GeoRegion-In-a-GeoRegion}"'},"​")],-1)),i("details",g,[i("summary",null,[s[4]||(s[4]=i("a",{id:"Base.in-Tuple{GeoRegion, GeoRegion}",href:"#Base.in-Tuple{GeoRegion, GeoRegion}"},[i("span",{class:"jlbinding"},"Base.in")],-1)),s[5]||(s[5]=n()),a(e,{type:"info",class:"jlObjectType jlMethod",text:"Method"})]),s[7]||(s[7]=t("",8)),a(e,{type:"info",class:"source-link",text:"source"},{default:o(()=>[...s[6]||(s[6]=[i("a",{href:"https://github.com/GeoRegionsEcosystem/GeoRegions.jl/blob/5b039db4afc150c0c3e68753b8ef289e3310d2c1/src/is/isin.jl#L64-L91",target:"_blank",rel:"noreferrer"},"source",-1)])]),_:1})]),s[42]||(s[42]=i("h2",{id:"On",tabindex:"-1"},[n("On "),i("a",{class:"header-anchor",href:"#On","aria-label":'Permalink to "On {#On}"'},"​")],-1)),s[43]||(s[43]=i("h3",{i
[…]

[…]
