### 2026-02-12 #8480 @hugoledoux
@openjournals/ese-eics This author doesn't respond anymore, I think we should put on hold (or stop it?)
