### 2026-02-13 #8480 @kthyng
@natgeo-wong Please get back to us within 1 week to let us know your timeline for finishing. At that point we will withdraw this submission on your behalf if we don't have a plan for finishing up.
